PDFBOX-4892: reduce memory usage, as suggested by valerybokov
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1893812 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/encryption/SecurityHandler.java
@@ -61,7 +61,7 @@ public abstract class SecurityHandler
 {
     private static final Log LOG = LogFactory.getLog(SecurityHandler.class);
 
-    private static final int DEFAULT_KEY_LENGTH = 40;
+    private static final short DEFAULT_KEY_LENGTH = 40;
 
     // see 7.6.2, page 58, PDF 32000-1:2008
     private static final byte[] AES_SALT = { (byte) 0x73, (byte) 0x41, (byte) 0x6c, (byte) 0x54 };
@@ -69,7 +69,7 @@ public abstract class SecurityHandler
     /**
      * The length in bits of the secret key used to encrypt the document. Will become private in 3.0.
      */
-    protected int keyLength = DEFAULT_KEY_LENGTH;
+    protected short keyLength = DEFAULT_KEY_LENGTH;
 
     /** The encryption key that will used to encrypt / decrypt. Will become private in 3.0. */
     protected byte[] encryptionKey;
@@ -676,7 +676,7 @@ public abstract class SecurityHandler
      */
     public void setKeyLength(int keyLen)
     {
-        this.keyLength = keyLen;
+        this.keyLength = (short) keyLen;
     }
 
     /**
